roomwise/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-Bcrypt
PyJWT
gunicorn
